1. Added FormColLinearSystem from flow branch. 2) Improved memory requirements for smoother by replacing the prolongation to restriction (of a patch) 3) Performed minor fixes in destructors  after checking with valgrind

+++ examples/Smoother/schwarz.cpp
@@ -11,7 +11,7 @@ using namespace mfem;
 
 void print(std::vector<int> const &input)
 {
-	for (int i = 0; i < input.size(); i++) {
+	for (int i = 0; i < (int)input.size(); i++) {
 		std::cout << input.at(i) << ' ';
 	}
 }
